Add img file deleting

+++ controllers/admin.js
@@ -1,6 +1,7 @@
+const ObjectId = require('mongoose').Types.ObjectId;
 const { validationResult } = require('express-validator');
 
-const ObjectId = require('mongoose').Types.ObjectId;
+const fileHelper = require('../util/file');
 
 const Product = require('../models/product');
 
@@ -136,6 +137,7 @@ exports.postEditProduct = (req, res, next) => {
       product.price = updatedPrice;
       product.description = updatedDescription;
       if (image) {
+        fileHelper.deleteFile(product.imageUrl);
         product.imageUrl = image.path;
       }
       return product.save().then((result) => {
@@ -168,8 +170,15 @@ exports.getProducts = (req, res, next) => {
 
 exports.postDeleteProduct = (req, res, next) => {
   const prodId = req.body.productId;
-  Product.deleteOne({ _id: prodId, userId: req.user._id })
-    // Product.findByIdAndRemove(prodId)
+  const DBprodId = new ObjectId(prodId);
+  Product.findById(DBprodId)
+    .then(product => {
+      if (!product) {
+        return next(new Error('Product not found!'));
+      }
+      fileHelper.deleteFile(product.imageUrl);
+      return Product.deleteOne({ _id: prodId, userId: req.user._id });
+    })
     .then((result) => {
       console.log('Product deleted from db!');
       res.redirect('products');
@@ -178,5 +187,6 @@ exports.postDeleteProduct = (req, res, next) => {
       const error = new Error(err);
       error.httpStatusCode = 500;
       return next(error);
-    });
+    })
+    .catch(err => next(err));
 };
